+++ Changelog.md
@@ -1,4 +1,7 @@
 # Changelog
 
+## Version (NEXT)
+* added processing of the tags attribute
+
 ## Version 0.1.0
 * initial release